feat: add overview of mocked SEP #5 to claim
