+++ src/components/Intro/index.tsx
@@ -1,4 +1,4 @@
-import { Grid, Typography, Box, Button, CircularProgress, SvgIcon } from '@mui/material'
+import { Grid, Typography, Box, Button, CircularProgress } from '@mui/material'
 import { useRouter } from 'next/router'
 import { formatEther } from 'ethers/lib/utils'
 import { BigNumber } from 'ethers'
@@ -20,19 +20,13 @@ import { useWallet } from '@/hooks/useWallet'
 import { isSafe } from '@/utils/wallet'
 import { InfoBox } from '@/components/InfoBox'
 import { useIsDelegationPending } from '@/hooks/usePendingDelegations'
-import { useIsDarkMode } from '@/hooks/useIsDarkMode'
 import { useMockSep5 } from '@/hooks/useMockSep5'
-import ClockIcon from '@/public/images/clock.svg'
-import { ExternalLink } from '@/components/ExternalLink'
+import { Sep5InfoBox } from '@/components/Sep5InfoBox'
 
 import css from './styles.module.css'
 
-const SEP5_URL =
-  'https://snapshot.org/#/safe.eth/proposal/0xb4765551b4814b592d02ce67de05527ac1d2b88a8c814c4346ecc0c947c9b941'
-
 export const Intro = (): ReactElement => {
   const router = useRouter()
-  const isDarkMode = useIsDarkMode()
   const isWrongChain = useIsWrongChain()
   const wallet = useWallet()
   const chainId = useChainId()
@@ -43,7 +37,6 @@ export const Intro = (): ReactElement => {
   const { total } = useTaggedAllocations()
 
   // TODO: Use real SEP #5 data
-  const fakeSep5ClaimDate = '01.01.1970'
   const sep5 = useMockSep5()
   const hasSep5Allocation = Number(sep5.allocation) > 0
   const isSep5Claimable = Number(sep5.claimable) > 0
@@ -112,33 +105,7 @@ export const Intro = (): ReactElement => {
 
           {hasSep5Allocation && (
             <Grid item xs={12}>
-              <InfoBox display="flex" alignItems="center" justifyContent="space-between">
-                <div>
-                  <Typography variant="body2" color="text.secondary">
-                    As a result of{' '}
-                    <ExternalLink href={SEP5_URL} icon={false}>
-                      SEP #5
-                    </ExternalLink>
-                    , you qualify for
-                  </Typography>
-                  <Typography fontWeight={700}>{formatAmount(sep5.allocation, 2)} SAFE</Typography>
-                </div>
-                <Typography
-                  variant="body2"
-                  sx={(theme) => ({
-                    py: 1,
-                    px: 2,
-                    bgcolor: isDarkMode ? theme.palette.primary.main : theme.palette.secondary.main,
-                    color: isDarkMode ? theme.palette.primary.contrastText : undefined,
-                    borderRadius: `${theme.shape.borderRadius}px`,
-                    display: 'flex',
-                    alignItems: 'center',
-                  })}
-                >
-                  <SvgIcon component={ClockIcon} inheritViewBox fontSize="inherit" sx={{ mr: '6px' }} />
-                  Until {fakeSep5ClaimDate}
-                </Typography>
-              </InfoBox>
+              <Sep5InfoBox />
             </Grid>
           )}
